+++ src/app/(retail-catalog)/cart/CartClient.tsx
@@ -1,14 +1,14 @@
 'use client';
 import RetailCartTable from '@/app/ui/product/RetailCartTable';
 import CartTable from '@/app/ui/product/CartTable';
-import { useAppSelector } from '@/lib/hooks';
+import { useIsB2B } from '@/lib/hooks';
 
 const CartClient = () => {
-  const { isAuthenticated, isB2b } = useAppSelector((state) => state.persistedAuthReducer);
+  const isB2b = useIsB2B();
 
   return (
     <main className="mx-auto min-h-[550px] w-full max-w-[1680px] bg-white px-4 pt-8 md:px-8 lg:px-[60px]">
-      {isAuthenticated && isB2b ? <CartTable /> : <RetailCartTable />}
+      {isB2b ? <CartTable /> : <RetailCartTable />}
       <div id="image" className="absolute z-50 hidden h-[200px] w-[200px]"></div>
     </main>
   );