+++ src/app/(retail-catalog)/history/HistoryClient.tsx
@@ -1,9 +1,9 @@
 'use client';
 import HistoryTable from '@/app/ui/product/HistoryTable';
-import { useAppSelector } from '@/lib/hooks';
+import { useIsB2B } from '@/lib/hooks';
 
 const HistoryClient = () => {
-  const { isB2b } = useAppSelector((state) => state.persistedAuthReducer);
+  const isB2b = useIsB2B();
 
   return (
     <main className="mx-auto min-h-[550px] w-full max-w-[1680px] bg-white px-4 pt-8 md:px-8 lg:px-[60px]">